requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
decorator==5.1.1
dynaconf==3.2.3
falcon==3.1.1
falcon-cors==1.1.7
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.10.0
gcsfs==2023.10.0
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
kiwisolver==1.4.5
matplotlib==3.8.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic_core==2.10.1
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
